fixed result compare

+++ trunk/sources/modules/agents/agents-impl/src/main/java/eu/scy/agents/search/searchresultenricher/SearchResultRanking.java
@@ -79,8 +79,7 @@ public class SearchResultRanking {
 				* similarityQuality + WEIGHT_RELEVANCE * relevanceQuality;
 	}
 
-	private double evaluateRelevance(List<ISearchResult> newResult,
-			List<ISearchResult> referenceResult) {
+	private double evaluateRelevance(List<ISearchResult> newResult, List<ISearchResult> referenceResult) {
 		// Compare the mean of the RELEVANCE_NUMBER first search results.
 		int n = RELEVANCE_NUMBER;
 		int min = Math.min(referenceResult.size(), newResult.size());
@@ -102,8 +101,7 @@ public class SearchResultRanking {
 		return (newMean + 1) / (refMean + 1);
 	}
 
-	private static int getIdenticalUriCount(List<ISearchResult> a,
-			List<ISearchResult> b) {
+	private static int getIdenticalUriCount(List<ISearchResult> a, List<ISearchResult> b) {
 		// get the number of identical eloUris
 		int identicalResults = 0;
 		for (ISearchResult bRes : b) {
@@ -124,70 +122,29 @@ public class SearchResultRanking {
 		// TODO full similiarity is overweighted...
 
 		// +1 to avoid division by zero
-		return (identicalUri + 1) / (minUri + 1);
+		return ((double)identicalUri + 1) / ((double)minUri + 1);
 	}
 
 	private static double evaluateLength(int newResult, int refResult) {
 		double quality;
 		if (newResult < refResult) {
 			// +1 to avoid division by zero
-			quality = ((refResult - newResult + 1) / (refResult + 1));
+			quality = (((double)refResult - newResult + 1) / ((double)refResult + 1));
 		} else {
-			quality = ((newResult - refResult + 1) / (newResult + 1));
+			quality = (((double)newResult - refResult + 1) / ((double)newResult + 1));
 		}
 
 		double penalty = 1.0;
 		if (newResult < SearchResultEnricherAgent.INFIMUM_GOOD_SEARCH_RESULT) {
 			// query has not enough results, so use a penalty
-			penalty = (SearchResultEnricherAgent.INFIMUM_GOOD_SEARCH_RESULT - newResult)
+			penalty = (SearchResultEnricherAgent.INFIMUM_GOOD_SEARCH_RESULT - (double)newResult)
 					/ SearchResultEnricherAgent.INFIMUM_GOOD_SEARCH_RESULT;
 		} else if (newResult > SearchResultEnricherAgent.SUPREMUM_GOOD_SEARCH_RESULT) {
 			// query has too many results, so use a penalty
-			penalty = (newResult - SearchResultEnricherAgent.INFIMUM_GOOD_SEARCH_RESULT)
+			penalty = ((double)newResult - SearchResultEnricherAgent.INFIMUM_GOOD_SEARCH_RESULT)
 					/ newResult;
 		}
 
-		// if(newResult < SearchResultEnricherAgent.INFIMUM_GOOD_SEARCH_RESULT)
-		// {
-		// // query has not enough results, so use a penalty
-		// penalty = (SearchResultEnricherAgent.INFIMUM_GOOD_SEARCH_RESULT -
-		// newResult) / SearchResultEnricherAgent.INFIMUM_GOOD_SEARCH_RESULT;
-		//
-		// if(newResult < refResult) {
-		// // ... and less than the reference - bad (+1 to avoid division by
-		// zero)
-		// quality = ((refResult - newResult + 1) / (refResult + 1));
-		// } else
-		// // ... but more than the reference - good (+1 to avoid division by
-		// zero)
-		// quality = ((newResult - refResult + 1) / (newResult + 1));
-		// }
-		//
-		// } else if(newResult >
-		// SearchResultEnricherAgent.SUPREMUM_GOOD_SEARCH_RESULT ) {
-		// // query has too many results, so use a penalty
-		// penalty = (newResult -
-		// SearchResultEnricherAgent.INFIMUM_GOOD_SEARCH_RESULT) / newResult;
-		//
-		// if(newResult > refResult) {
-		// // ... and more than the reference - bad
-		// quality -= ((newResult - refResult) / (newResult));
-		// } else if(newResult > refResult) {
-		// // ... but less than the reference - good
-		// quality += ((refResult - newResult) / (refResult));
-		// }
-		// }
-		// else {
-		// // result number is ok
-		//
-		// if(newResult > refResult) {
-		// // ... and more than the reference - good
-		// quality += ((newResult - refResult) / (newResult));
-		// } else if(newResult > refResult) {
-		// // ... but less than the reference - bad
-		// quality -= ((refResult - newResult) / (refResult));
-		// }
-		// }
 		return quality * penalty;
 	}
 }
